added high score save/load
high scores are saved in a game save file, loaded when game opens (or
default used if one doesn't exist), and saved when player gets a new
high score and enters initials. also modified camera to not be locked in
z with player to more easily see platform levels when jumping.