added sounds effects and music
added background music and sound effects.

all sound files in Content\SideScrollerBP\sounds

changing played sound is done by saving over existing wav file and
reimporting it in the editor (just chose some music and sound effects
for now, they can be changed at any time)

adjusting volume of individual sounds is done in the assosciated sound
cue object.

adjusting volume of all sound effects or all background music is done
with the two floats named for that in the character variables (not
dynamic, won't work during game, unless mute is toggled with M or a new
level starts, could put in dynamic changing of overall volume, but it
seems unecessary since user can change volume of the program on their
desktop and get the same effect, but perhaps sliders for msuic and
effects to control independantly could be an option in some menu
somewhere, i still think that's too much for the scope of the game,
unless we have lots of time left)

adjusting sound fall-off distance is done in the attentuation settings,
currently only bouncer has this, of the audio component of the class
that's making the sound

not a huge amount of testing done with sounds, if you hear something you
don't expect that might not be intended.

dissallowed double jumps, if it was already used after taking damage
from soniq and spike from any direction (not sure on desired gameplay
here actually, should double jump reset when taking damage from above
and getting knocked upward?)