grader shortcut, and gameplay changes
changed vort enemies animation based on walking speed (which is based on
difficulty).

added more grader shortcuts and they are saved, and moved them into the
play menu, since that's where they're relevant.

changed it to double jump refreshes when hit by any enemy (as opposed to
not refreshing if enemy did damage, could revert this though, but i
think it makes the game feel better.