more custcene integration and item/enemy spawns on elevators
muted all other sounds when cutscene is playing.

mute button works to mute custcene sound.

if game was muted when it was closed last, cutscene is muted when game
is opened again.

put custcene music in a cue object and assigned it to music class so it
gets muted w/ the mute button.

user can click during cutscene to end it early (so they're not forced to
watch it every time if they don't want to).

Note that making a click do that means when the play in editor is
running, we usually need to click once in the game to get control to
that window, and doing that stops the cutscene, but in the packaged game
that window is automatically in focus so that won't be a problem for the
final executable.

added coins and enemies to elevators.